backend/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
python-multipart
python-dotenv
scikit-learn
numpy
